requirements.txt
appdirs==1.4.3
audioread==2.1.4
click==6.7
ez-setup==0.9
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysql-python==1.2.5
packaging==16.8
passlib==1.7.1
pyacoustid==1.1.5
pyparsing==2.2.0
requests==2.14.2
six==1.10.0
Werkzeug==0.12.2
WTForms==2.1
